requirements.txt
psycopg2-binary==2.9.9
fastapi==0.140.0
uvicorn==0.34.0
pydantic==2.10.4
pytest==8.4.2
httpx==0.28.1